requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.12.0
django-model-utils==4.1.1
Pillow==8.4.0
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.13.1
uWSGI>=2.0.18,<2.1
psycopg2>=2.8.6,<2.9